refactor: Navigator -> Launcher 네이밍 변경
