+++ core/launcher/src/main/java/com/acon/acon/core/launcher/AppLauncher.kt
@@ -1,0 +1,7 @@
+package com.acon.acon.core.launcher
+
+import android.content.Context
+
+interface AppLauncher {
+    fun restartApp(context: Context)
+}
