--- core/navigation/src/main/java/com/acon/acon/core/navigation/navigator/AppNavigator.kt
@@ -1,8 +1,0 @@
-package com.acon.acon.core.navigation.navigator
-
-import android.content.Context
-
-interface AppNavigator {
-
-    fun restartApp(context: Context)
-}
